Use CSpell config file
Use a CSpell config file rather than inline comments to ignore words.
This allows to ignore words at the project level rather than for
each individual source file. The latter could be useful for words that
are really specific/private to a single source file, but for such a
small project, this is not going to be an issue.